Split PlayerMgmt controllers into Player and Group
