+++ src/Nether.Web/Features/PlayerManagement/Controllers/PlayerController.cs
@@ -139,72 +139,5 @@ namespace Nether.Web.Features.PlayerManagement
             // Return result
             return NoContent();
         }
-
-        /// <summary>
-        /// Gets the list of groups current player belongs to.
-        /// </summary>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupListResponseModel))]
-        [HttpGet("player/groups")]
-        public async Task<ActionResult> GetPlayerGroups()
-        {
-            var gamertag = User.GetGamerTag();
-            var groups = await _store.GetPlayersGroupsAsync(gamertag);
-
-            return Ok(GroupListResponseModel.FromGroups(groups));
-        }
-
-        /// <summary>
-        /// Adds currently logged in player to a group.
-        /// </summary>
-        /// <param name="groupName">Group name.</param>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        [HttpPut("player/groups/{groupName}")]
-        public async Task<ActionResult> AddCurrentPlayerToGroup(string groupName)
-        {
-            var gamertag = User.GetGamerTag();
-            Group group = await _store.GetGroupDetailsAsync(groupName);
-            if (group == null)
-            {
-                _logger.LogWarning("group '{0}' not found", groupName);
-                return BadRequest();
-            }
-
-            Player player = await _store.GetPlayerDetailsByGamertagAsync(gamertag);
-            if (player == null)
-            {
-                _logger.LogWarning("player '{0}' not found", gamertag);
-                return BadRequest();
-            }
-
-            await _store.AddPlayerToGroupAsync(group, player);
-
-            return NoContent();
-
-        }
-
-        /// <summary>
-        /// Gets the members of the group as player objects.
-        /// </summary>
-        /// <param name="groupName">Name of the group</param>
-        /// <returns></returns>
-        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(GroupMemberResponseModel))]
-        [HttpGet("groups/{groupName}/players")]
-        public async Task<ActionResult> GetGroupPlayers(string groupName)
-        {
-            // Call data store
-            List<string> gamertags = await _store.GetGroupPlayersAsync(groupName);
-
-            // Format response model
-            var resultModel = new GroupMemberResponseModel
-            {
-                Gamertags = gamertags
-            };
-
-            // Return result
-            return Ok(resultModel);
-        }
     }
 }
